+++ es/resource_elasticsearch_xpack_role_test.go
@@ -16,6 +16,7 @@ import (
 )
 
 func TestAccElasticsearchXpackRole(t *testing.T) {
+
 	provider := Provider().(*schema.Provider)
 	err := provider.Configure(&terraform.ResourceConfig{})
 	if err != nil {
@@ -187,6 +188,7 @@ func testAccRoleResource(resourceName string) string {
 	}
 	`, resourceName)
 }
+
 func testAccRoleResource_Updated(resourceName string) string {
 	return fmt.Sprintf(`
 	resource "elasticsearch_xpack_role" "test" {